Split actual data type classes from TexuteToNativeTensor
