+++ Packages/com.github.asus4.tflite.common/Runtime/TextureToNativeTensorDataTypes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b977be6022e0143d6bf6a22f1d02934b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
